Feat : 차량 중복 예약 방지 완료

+++ src/main/java/com/douzone/server/service/VehicleService.java
@@ -41,7 +41,6 @@ public class VehicleService {
 	@Transactional
 	public ResponseDTO registerByVehicleReservation(VehicleParseDTO vehicleParseDTO, Long empId) {
 		log.info(METHOD_NAME + "- registerByVehicleReservation");
-
 		VehicleReservationDTO vehicleReservationDTO = VehicleReservationDTO.builder()
 				.vehicleId(vehicleParseDTO.getVehicleId())
 				.empId(vehicleParseDTO.getEmpId())
@@ -50,13 +49,12 @@ public class VehicleService {
 				.startedAt(parsing(vehicleParseDTO.getStartedAt()))
 				.endedAt(parsing(vehicleParseDTO.getEndedAt()))
 				.build();
-
 		return Optional.of(new ResponseDTO())
 				.filter(u -> empId > 0)
 				.map(res -> {
 					vehicleReservationRepository.save(
 							VehicleReservation.builder()
-									.vehicle(Vehicle.builder().id((long) vehicleReservationDTO.getVehicleId()).build())
+									.vehicle(Vehicle.builder().id(vehicleReservationDTO.getVehicleId()).build())
 									.employee(Employee.builder().id(empId).build())
 									.reason(vehicleReservationDTO.getReason())
 									.title(vehicleReservationDTO.getTitle())
@@ -71,7 +69,6 @@ public class VehicleService {
 	@Transactional
 	public ResponseDTO registerByVehicleBookmark(Long empId, Long vId) {
 		log.info(METHOD_NAME + "- registerByVehicleBookmark");
-
 		return Optional.of(new ResponseDTO())
 				.filter(u -> empId > 0)
 				.filter(v -> vId > 0)
@@ -89,7 +86,6 @@ public class VehicleService {
 	@Transactional(readOnly = true)
 	public ResponseDTO findByAllReservation() {
 		log.info(METHOD_NAME + "- findByAllReservation");
-
 		return Optional.of(new ResponseDTO())
 				.map(u -> ResponseDTO.of(HttpStatus.OK, SUCCESS_VEHICLE_FIND_ALL, vehicleRepository.findByAllReservation()))
 				.orElseGet(() -> ResponseDTO.fail(HttpStatus.BAD_REQUEST, FAIL_VEHICLE_FIND_ALL + FAIL_EXIST_RESULT));
@@ -98,7 +94,6 @@ public class VehicleService {
 	@Transactional(readOnly = true)
 	public ResponseDTO findByPaginationReservation(int pageNum) {
 		log.info(METHOD_NAME + "- findByPaginationReservation");
-
 		return Optional.of(new ResponseDTO())
 				.map(u -> (pageNum < 0) ?
 						ResponseDTO.fail(HttpStatus.BAD_REQUEST, FAIL_VEHICLE_FIND_ALL + FAIL_REQUEST_PARAMETER) :
@@ -109,7 +104,6 @@ public class VehicleService {
 	@Transactional(readOnly = true)
 	public ResponseDTO findByAllNotReservation() {
 		log.info(METHOD_NAME + "- findByAllNotReservation");
-
 		return Optional.of(new ResponseDTO())
 				.map(u -> ResponseDTO.of(HttpStatus.OK, SUCCESS_VEHICLE_FIND_NONE, vehicleRepository.findByAllNotReservation(now().plusHours(1L))))
 				.orElseGet(() -> ResponseDTO.fail(HttpStatus.BAD_REQUEST, FAIL_VEHICLE_FIND_NONE + FAIL_EXIST_RESULT));
@@ -128,7 +122,6 @@ public class VehicleService {
 	@Transactional(readOnly = true)
 	public ResponseDTO findByDateTimeReservation(String start, String end) {
 		log.info(METHOD_NAME + "- findByDateTimeReservation");
-
 		return Optional.of(new ResponseDTO())
 				.map(u -> (start == null || start.equals("") || end == null || end.equals("")) ?
 						ResponseDTO.fail(HttpStatus.BAD_REQUEST, FAIL_VEHICLE_FIND_DATE + FAIL_REQUEST_PARAMETER) :
@@ -141,7 +134,6 @@ public class VehicleService {
 	@Transactional(readOnly = true)
 	public ResponseDTO findByMyPastReservation(Long id) {
 		log.info(METHOD_NAME + "- findByMyPastReservation");
-
 		return Optional.of(new ResponseDTO())
 				.map(u -> (id != null) ?
 						ResponseDTO.of(HttpStatus.OK, SUCCESS_VEHICLE_BEFORE, vehicleRepository.findByMyPastReservation(id)) :
@@ -152,7 +144,6 @@ public class VehicleService {
 	@Transactional(readOnly = true)
 	public ResponseDTO findByMyCurrentReservation(Long id) {
 		log.info(METHOD_NAME + "- findByMyCurrentReservation");
-
 		return Optional.of(new ResponseDTO())
 				.map(u -> (id != null) ?
 						ResponseDTO.of(HttpStatus.OK, SUCCESS_VEHICLE_AFTER, vehicleRepository.findByMyCurrentReservation(id)) :
@@ -163,7 +154,6 @@ public class VehicleService {
 	@Transactional(readOnly = true)
 	public ResponseDTO weekMostReservedVehicle() {
 		log.info(METHOD_NAME + "- weekMostReservedVehicle");
-
 		return Optional.of(new ResponseDTO())
 				.map(u -> ResponseDTO.of(HttpStatus.OK, SUCCESS_VEHICLE_BEST_WEEK, vehicleRepository.weekMostReservedVehicle(now().minusDays(7L))))
 				.orElseGet(() -> ResponseDTO.fail(HttpStatus.BAD_REQUEST, FAIL_VEHICLE_BEST_WEEK + FAIL_EXIST_RESULT));
@@ -172,7 +162,6 @@ public class VehicleService {
 	@Transactional(readOnly = true)
 	public ResponseDTO weekMostReservedTime() {
 		log.info(METHOD_NAME + "- weekMostReservedTime");
-
 		return Optional.of(new ResponseDTO()).map(u -> {
 			Map<String, Integer> map = new HashMap<>();
 			List<VehicleWeekTimeDTO> result = vehicleRepository.weekMostReservedTime(now().minusDays(7L), now());
@@ -191,7 +180,6 @@ public class VehicleService {
 	@Transactional(readOnly = true)
 	public ResponseDTO findByRecentReservedVehicle() {
 		log.info(METHOD_NAME + "- findByRecentReservedVehicle");
-
 		return Optional.of(new ResponseDTO())
 				.map(u -> ResponseDTO.of(HttpStatus.OK, SUCCESS_VEHICLE_RECENT, vehicleRepository.findByRecentReservedVehicle()))
 				.orElseGet(() -> ResponseDTO.fail(HttpStatus.BAD_REQUEST, FAIL_VEHICLE_RECENT + FAIL_EXIST_RESULT));
@@ -200,7 +188,6 @@ public class VehicleService {
 	@Transactional(readOnly = true)
 	public ResponseDTO findByMyBookMarkVehicle(String empNo) {
 		log.info(METHOD_NAME + "- findMarkVehicle");
-
 		return Optional.of(new ResponseDTO())
 				.map(u -> ResponseDTO.of(HttpStatus.OK, SUCCESS_VEHICLE_FIND_MARK, vehicleBookmarkRepository.findByMyBookMarkVehicle(empNo)))
 				.orElseGet(() -> ResponseDTO.fail(HttpStatus.BAD_REQUEST, FAIL_VEHICLE_FIND_MARK + FAIL_EXIST_RESULT));
@@ -209,7 +196,6 @@ public class VehicleService {
 	@Transactional
 	public ResponseDTO selectByBookMarkTop3Vehicle() {
 		log.info(METHOD_NAME + "- selectByBookMarkTop3Vehicle");
-
 		return Optional.of(new ResponseDTO())
 				.map(u -> ResponseDTO.of(HttpStatus.OK, SUCCESS_VEHICLE_BEST_MARK, vehicleBookmarkRepository.selectByBookMarkTop3Vehicle(PageRequest.of(0, 3))))
 				.orElseGet(() -> ResponseDTO.fail(HttpStatus.BAD_REQUEST, FAIL_VEHICLE_BEST_MARK + FAIL_EXIST_RESULT));
@@ -218,7 +204,6 @@ public class VehicleService {
 	@Transactional
 	public ResponseDTO updateByMyReservation(VehicleParseDTO vehicleParseDTO, Long id) {
 		log.info(METHOD_NAME + "- updateByMyReservation");
-
 		VehicleReqDTO vehicleReqDTO = VehicleReqDTO.builder()
 				.id(vehicleParseDTO.getId())
 				.vehicleId(vehicleParseDTO.getVehicleId())
@@ -243,7 +228,6 @@ public class VehicleService {
 	@Transactional
 	public ResponseDTO deleteByMyReservation(Long id, Long empId) {
 		log.info(METHOD_NAME + "- deleteByMyReservation");
-
 		return Optional.of(new ResponseDTO())
 				.filter(u -> (id != null))
 				.map(u -> vehicleReservationRepository.findById(id))
@@ -260,7 +244,6 @@ public class VehicleService {
 	@Transactional
 	public ResponseDTO deleteByMyBookMark(Long id) {
 		log.info(METHOD_NAME + "- deleteByMyBookMark");
-
 		return Optional.of(new ResponseDTO())
 				.filter(u -> (id != null))
 				.map(v -> {
@@ -274,7 +257,6 @@ public class VehicleService {
 	@Transactional
 	public ResponseDTO selectByVehicleReservation(Long id) {
 		log.info(METHOD_NAME + "- selectByVehicleReservation");
-
 		return Optional.of(new ResponseDTO())
 				.filter(u -> (id != null))
 				.map(v -> vehicleRepository.selectByVehicleReservation(id))
@@ -287,7 +269,6 @@ public class VehicleService {
 	@Transactional
 	public ResponseDTO soonAndIngReservationMyTime(Long empId, int code) {
 		log.info(METHOD_NAME + "- soonReservationMyTime");
-
 		return Optional.of(new ResponseDTO())
 				.filter(u -> (empId > 0))
 				.map(v -> code == 0 ?
@@ -314,7 +295,6 @@ public class VehicleService {
 	@Transactional
 	public ResponseDTO earlyReturnOfVehicle(Long id, Long empId) {
 		log.info(METHOD_NAME + "- earlyReturnOfVehicle");
-
 		return Optional.of(new ResponseDTO())
 				.filter(u -> (id > 0))
 				.map(u -> vehicleReservationRepository.findById(id))
@@ -328,7 +308,6 @@ public class VehicleService {
 							.startedAt(v.get().getStartedAt())
 							.endedAt(now())
 							.build());
-
 					return ResponseDTO.of(HttpStatus.OK, SUCCESS_MOVE_UP_RESERVATION);
 				}).orElseGet(() -> ResponseDTO.fail(HttpStatus.BAD_REQUEST, FAIL_MOVE_UP_RESERVATION));
 	}
@@ -336,7 +315,6 @@ public class VehicleService {
 	@Transactional(readOnly = true)
 	public ResponseDTO selectByVariousColumns(VehicleSearchDTO vehicleSearchDTO) {
 		log.info(METHOD_NAME + "- selectByVariousColumns");
-
 		return Optional.of(new ResponseDTO())
 				.map(v -> vehicleQueryDSL.selectByVariousColumns(vehicleSearchDTO))
 				.map(v -> v.isEmpty() ?
